Move SimpleCodeBlock to ui-patterns.
