+++ apps/studio/components/interfaces/Docs/CodeSnippet.tsx
@@ -1,5 +1,5 @@
 import { useParams } from 'common'
-import { SimpleCodeBlock } from 'ui'
+import { SimpleCodeBlock } from 'ui-patterns/SimpleCodeBlock'
 
 import { useSendEventMutation } from '@/data/telemetry/send-event-mutation'
 import { useSelectedOrganizationQuery } from '@/hooks/misc/useSelectedOrganization'
