+++ apps/studio/components/interfaces/TableGridEditor/ViewEntityAutofixSecurityModal.tsx
@@ -1,14 +1,15 @@
 import { useQueryClient } from '@tanstack/react-query'
 import { toast } from 'sonner'
-
-import { useViewDefinitionQuery } from 'data/database/view-definition-query'
-import { lintKeys } from 'data/lint/keys'
-import { useExecuteSqlMutation } from 'data/sql/execute-sql-mutation'
-import { Entity, isViewLike } from 'data/table-editor/table-editor-types'
-import { useSelectedProjectQuery } from 'hooks/misc/useSelectedProject'
-import { ScrollArea, SimpleCodeBlock } from 'ui'
-import { GenericSkeletonLoader } from 'ui-patterns'
+import { ScrollArea } from 'ui'
 import ConfirmationModal from 'ui-patterns/Dialogs/ConfirmationModal'
+import { GenericSkeletonLoader } from 'ui-patterns/ShimmeringLoader'
+import { SimpleCodeBlock } from 'ui-patterns/SimpleCodeBlock'
+
+import { useViewDefinitionQuery } from '@/data/database/view-definition-query'
+import { lintKeys } from '@/data/lint/keys'
+import { useExecuteSqlMutation } from '@/data/sql/execute-sql-mutation'
+import { Entity, isViewLike } from '@/data/table-editor/table-editor-types'
+import { useSelectedProjectQuery } from '@/hooks/misc/useSelectedProject'
 
 interface ViewEntityAutofixSecurityModalProps {
   table: Entity
