+++ packages/ui/index.tsx
@@ -10,7 +10,6 @@ export * from './src/components/Tabs'
 export * from './src/components/Alert'
 export * from './src/components/Accordion'
 export * from './src/components/Collapsible'
-export * from './src/components/SimpleCodeBlock'
 
 // NAV
