+++ packages/ui/package.json
@@ -60,8 +60,6 @@
     "input-otp": "^1.2.3",
     "lodash": "catalog:",
     "lucide-react": "^0.436.0",
-    "next-themes": "^0.3.0",
-    "prism-react-renderer": "^2.3.1",
     "prop-types": "^15.7.2",
     "react": "catalog:",
     "react-accessible-treeview": "^2.11.2",
